back/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
databases[postgresql]==0.9.0
sqlalchemy==2.0.36
asyncpg==0.30.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pydantic-settings==2.6.1
alembic==1.14.0
psycopg2-binary==2.9.10
httpx==0.28.1
python-dotenv==1.0.1
pandas==2.2.0
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0